Separate most of the code to a package
Almost everything (except train.py) regarding either the model or the dataset has been moved to emotions_classifier package.
This greatly reduces amount of files exposed to the user when trying to launch the project.
