+++ main.py
@@ -1,20 +1,20 @@
 from sklearn.model_selection import train_test_split
-from dataset_utils import unzip_nested_dataset, load_ravdess_data, download_dataset_from_gdrive, RAVDESS_DOWNLOAD_URL,\
-    RAVDESS_PATH
-from RAVDESSDataset import RAVDESSDataset
-from misc import load_model_for_inference
+from emotions_classifier.utils.dataset_utils import unzip_nested_dataset, load_ravdess_data, download_dataset_from_gdrive, RAVDESS_DOWNLOAD_URL
+from emotions_classifier import RAVDESS_ZIP_PATH, RAVDESS_DATASET_DIR
+from emotions_classifier.RAVDESSDataset import RAVDESSDataset
+from emotions_classifier.utils.misc import load_model_for_inference
 
 MODEL_PATH = "EmotionsClassifier.pth"
 MODEL_URL = "https://drive.google.com/uc?id=1rEBqU3geg2V4_W9QGY-nZGWWdNXd4zrq"
 
 
 def inference():
-    download_dataset_from_gdrive(RAVDESS_DOWNLOAD_URL, RAVDESS_PATH)
+    download_dataset_from_gdrive(RAVDESS_DOWNLOAD_URL, RAVDESS_ZIP_PATH)
     class_names = [
         "Neutral", "Calm", "Happy", "Sad", "Angry", "Fearful", "Disgust", "Surprised"
     ]
-    unzip_nested_dataset("data/ravdess.zip", "ravdess_data")
-    file_paths, labels = load_ravdess_data("ravdess_data", audio_type="speech")
+    unzip_nested_dataset(RAVDESS_ZIP_PATH, RAVDESS_DATASET_DIR)
+    file_paths, labels = load_ravdess_data(RAVDESS_DATASET_DIR, audio_type="speech")
 
     # Split the dataset in the exact same way as with training
     train_files, _, train_labels, _ = train_test_split(
